# hackathon/openapi.yaml
openapi: 3.0.3
info:
  title: "GLIA Holographic Memory API"
  description: "Cloud-Native Epistemic Memory for AI Agents. Use this API to recall historical project context, past architectural decisions, or bugs based on pattern resonance, rather than exact keywords."
  version: 1.0.0
servers:
  - url: "https://glia-memory-api-322362297536.us-central1.run.app"
    description: "Production Cloud Run Deployment"
paths:
  /recall:
    post:
      operationId: recall_memory
      summary: "Retrieve historical context from holographic memory"
      description: "Given a query (e.g., a code snippet, an error message, or a technical concept), this endpoint projects the query into the holographic substrate and returns the resonating historical context. Use this whenever the agent needs to understand the history or architecture of a project before making a decision."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: "The technical concept, error log, or code snippet to search for."
                top_k:
                  type: integer
                  description: "Maximum number of resonant nodes to return."
                  default: 5
              required:
                - query
      responses:
        '200':
          description: "Successful retrieval"
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: string
                    description: "A structured cognitive map explaining the historical context related to the query."
                  nodes:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
                    description: "The raw IDs and resonance scores of the activated glyphs."
  /learn:
    post:
      operationId: learn_memory
      summary: "Teach a new concept to the holographic memory"
      description: "Use this to save important decisions, bug fixes, or architecture rules to the long-term memory. The content will be distilled by Gemini into concepts and saved holographically."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: "The text containing the knowledge to save (e.g. Incident 402: Never use JSON.stringify in logs)."
                source:
                  type: string
                  description: "The origin of this knowledge (e.g. GitLab MR 123, User command)."
              required:
                - content
      responses:
        '200':
          description: "Knowledge successfully learned"
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  concepts:
                    type: array
                    items:
                      type: string
  /stats:
    get:
      operationId: get_stats
      summary: "Get memory statistics"
      description: "Returns the number of glyphs, regions, and the dimension of the holographic substrate."
      responses:
        '200':
          description: "Successful response"
          content:
            application/json:
              schema:
                type: object
                properties:
                  dimension:
                    type: integer
                  regions:
                    type: integer
                  glyphs:
                    type: integer
                  total_capacity:
                    type: integer